build(chore): restructure source files
