+++ src/background.js
@@ -1,4 +1,4 @@
-import { createMenu, addMenuEventListeners } from './js/menu.js'
+import { createMenu, addMenuEventListeners } from './components/menu.js'
 
 chrome.runtime.onInstalled.addListener((_reason) => {
   createMenu()
